refactor: create a standalone opt out partial template
