+++ header/main.scss
@@ -16,7 +16,7 @@ $o-header-image-base-url: 'https://next-geebee.ft.com/image';
 	}
 
 	// add next opt-out buttons
-	.o-header__drawer-menu-opt-out {
+	.o-header__drawer-menu-button {
 		@include oButtons('medium', 'standout');
 		display: block;
 		margin: $_o-header-drawer-padding-x;
