+++ opt-out/template.html
@@ -1,0 +1,9 @@
+{{#if @root.flags.optInOut}}
+	{{#unless @root.forceOptInDevice}}
+		<a class="{{classname}} js-optout-link" data-trackable="opt-out" href="/opt-out-confirm">
+			Return to <span class="{{classname}}-extended">the old</span> FT.com
+		</a>
+	{{/unless}}
+	<a class="{{classname}} js-webapp-link" data-trackable="visit-webapp" href="https://app.ft.com" hidden>Open in app</a>
+	<a class="{{classname}} js-android-link" data-trackable="open-android-app" href="http://www.ft.com/androidapp" hidden>Open in app</a>
+{{/if}}
