+++ welcome-message/message.html
@@ -11,18 +11,6 @@
 		<a class="n-welcome-message__button n-welcome-message__button--feedback" data-trackable="nps-feedback" href="/nps-feedback">
 			<span class="n-welcome-message__button-extended">Provide</span> Feedback
 		</a>
-		{{#if @root.flags.optInOut}}
-			{{#unless @root.forceOptInDevice}}
-			<a class="n-welcome-message__button n-welcome-message__button--opt-out js-optout-link" data-trackable="opt-out" href="/opt-out-confirm">
-				Return to <span class="n-welcome-message__button-extended">the old</span> FT.com
-			</a>
-			{{/unless}}
-			<a class="n-welcome-message__button n-welcome-message__button--webapp js-webapp-link" data-trackable="visit-webapp" href="https://app.ft.com" hidden>
-				Open in app
-			</a>
-			<a class="n-welcome-message__button n-welcome-message__button--android js-android-link" data-trackable="open-android-app" href="http://www.ft.com/androidapp" hidden>
-				Open in app
-			</a>
-		{{/if}}
+		{{>n-ui/opt-out/template classname='n-welcome-message__button'}}
 	</div>
 </div>
